Refactoring
Improved comments, moved MF2-related code into a new file, renamed
some functions
