+++ package.json
@@ -46,7 +46,7 @@
   "scripts": {
     "build": "yarn clean && cross-env NODE_ENV=production && yarn tsc",
     "clean": "yarn clean:build && yarn clean:examples",
-    "clean:build": "rm -rf lib plugin appWith* Dynamic* I18n* index _context loadNa* setLang* Trans* useT* withT* getP* getC* *.d.ts getT transC* wrapT* types formatElements",
+    "clean:build": "rm -rf lib plugin appWith* Dynamic* I18n* index _context loadNa* setLang* Trans* useT* withT* getP* getC* *.d.ts getT transC* wrapT* types formatElements messageFormatHelpers",
     "clean:examples": "rm -rf examples/**/.next && rm -rf examples/**/node_modules && rm -rf examples/**/yarn.lock",
     "example": "yarn example:complex",
     "example:basic": "yarn build && cd examples/basic && yarn && yarn dev",
@@ -80,7 +80,7 @@
     "jest": "27.3.1",
     "next": "12.3.1",
     "prettier": "2.4.1",
-    "pretty-quick": "3.1.2",
+    "pretty-quick": "4.0.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "supertest": "6.1.6",
